backend/requirements.txt
fastapi==0.104.1
mangum==0.17.0
python-multipart==0.0.6
google-generativeai==0.3.2
